// src/protos/api_pb.d.ts
export enum Entity {
  ORGANIZATION = 0,
  SCHOOL = 1,
  CLASS = 2,
  USER = 3,
  ROLE = 4,
  PROGRAM = 5,
  UNKNOWN = 15,
}